pyproject.toml
[project]
name = "aicallcenter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[websockets]>=0.129.0",
    "langgraph>=1.0.9",
    "loguru>=0.7.3",
    "openrouter>=0.6.0",
    "psycopg2-binary>=2.9.11",
    "pyannote-audio>=4.0.4",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.4",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
